requirements.txt
mss~=9.0.1
pynput~=1.7.7
transformers~=4.41.2
numpy~=1.22.0
opencv-python~=4.10.0.84
pillow~=10.3.0
requests~=2.32.3
pyinstaller~=6.9.0